extracts listener setup to hook
