+++ src/providers/fractal/FractalProvider.tsx
@@ -1,11 +1,12 @@
-import { ReactNode, useEffect, useMemo, useReducer } from 'react';
+import { ReactNode, useMemo, useReducer } from 'react';
 
 import { initialState } from './constants';
 import { FractalAction } from './constants/enums';
 import { useDAOLegacy } from './hooks/useDAOLegacy';
 import { FractalContext } from './hooks/useFractal';
-import { ACRoleListener, FractalActions, FractalDAO, IDaoLegacy } from './types';
+import { FractalActions, FractalDAO, IDaoLegacy } from './types';
 import { useModuleTypes } from './hooks/useModuleTypes';
+import { useModuleListeners } from './hooks/useModuleListeners';
 
 const initializeState = (_initialState: FractalDAO) => {
   return _initialState;
@@ -26,6 +27,9 @@ export const reducer = (state: FractalDAO, action: FractalActions): FractalDAO =
   }
 };
 
+/**
+ * Uses Context API to provider DAO information to app
+ */
 export function FractalProvider({ children }: { children: ReactNode }) {
   const [dao, dispatch] = useReducer(reducer, initialState, initializeState);
   const daoLegacy: IDaoLegacy = useDAOLegacy(dao.daoAddress);
@@ -33,47 +37,7 @@ export function FractalProvider({ children }: { children: ReactNode }) {
   const { timelockModule, treasuryModule, tokenVotingGovernanceModule, claimingContractModule } =
     useModuleTypes(dao.moduleAddresses);
 
-  useEffect(() => {
-    if (!dao.daoContract || !dao.accessControlContract) {
-      return;
-    }
-    const addFilter = dao.accessControlContract.filters.ActionRoleAdded();
-
-    const addRoleEventListener: ACRoleListener = (target: string) => {
-      if (target === dao.daoContract!.address) {
-        return;
-      }
-      const addresses = dao.moduleAddresses || [];
-      if (!addresses.includes(target)) {
-        dispatch({
-          type: FractalAction.UPDATE_MODULE,
-          payload: [...addresses, target],
-        });
-      }
-    };
-    const removefilter = dao.accessControlContract.filters.ActionRoleRemoved();
-
-    const removeRoleListener: ACRoleListener = (target: string) => {
-      if (target === dao.daoContract!.address) {
-        return;
-      }
-      const addresses = dao.moduleAddresses || [];
-      if (!addresses?.includes(target)) {
-        dispatch({
-          type: FractalAction.UPDATE_MODULE,
-          payload: [...addresses, target],
-        });
-      }
-    };
-
-    dao.accessControlContract.on(addFilter, addRoleEventListener);
-    dao.accessControlContract.on(removefilter, removeRoleListener);
-
-    return () => {
-      dao.accessControlContract!.off(addFilter, addRoleEventListener);
-      dao.accessControlContract!.off(removefilter, removeRoleListener);
-    };
-  }, [dao.daoContract, dao.accessControlContract, dao.moduleAddresses]);
+  useModuleListeners(dao, dispatch);
 
   const value = useMemo(
     () => ({
